+++ pychron/processing/tasks/analysis_edit/analysis_edit_task.py
@@ -245,10 +245,13 @@ class AnalysisEditTask(BaseBrowserTask):
                 if self.browser_model.use_workspace:
                     ans = self.workspace.make_analyses(records)
                 else:
+                    print 'records', records
                     ans = self.manager.make_analyses(records, calculate_age=True, load_aux=True)
+                    print 'aa', ans
                 self._open_recall_editors(ans)
         else:
             ans = self.manager.make_analyses(records, use_cache=False, calculate_age=True, load_aux=True)
+            print 'bb', ans
             self._open_recall_editors(ans)
 
     def _selector_dclick(self, new):
@@ -292,8 +295,8 @@ class AnalysisEditTask(BaseBrowserTask):
                 except AttributeError:
                     pass
 
-            ed = self.editor_area.editors[-1]
-            self.editor_area.activate_editor(ed)
+            # ed = self.editor_area.editors[-1]
+            # self.editor_area.activate_editor(ed)
         else:
             self.warning('could not load records')